Can now select cards for matching
Added enough functionality to almost completely finish a one player game.  The player can select three cards; they become unhighlighted if there is no match.  If there is a match, the player's score increases, and three new cards replace the matched ones.  Also added code "under the hood"; made methods that update the card buttons based off of the Board array that holds the board cards, and created initializers for the Player objects in the GameRuler constructor.  Score is saved to the player objects (which are stored within GameRuler), and score updates appropriately on the GUI.