Replace matched cards
Added a method that replaces three matched cards on the game board with three new cards from the deck.  I created an ArrayList in GameWindow that stores the selected cards; this is used as an input to tell the method which of the 12 cards should be replaced.  Also fixed a bug in the method that checks for all solutions on the board; it was referencing an index that was out of bounds.